requirements.txt
numpy
scipy
python-telegram-bot>=21.0
